dataprep/requirements.txt
# Dataset preparation dependencies
# Works with Python 3.8+ - no MM* dependencies!

# HTTP downloads with progress
requests>=2.28.0,<3.0.0

# Core utilities
numpy>=1.24.0,<3.0.0
Pillow>=10.0.0,<13.0.0
tqdm>=4.66.0,<5.0.0
pyyaml>=6.0,<7.0.0

# Optional: For faster json parsing
orjson>=3.9.0,<4.0.0

# Image processing
opencv-python>=4.9.0,<5.0.0
